+++ src/Blockchain/Blockchain.py
@@ -1,4 +1,5 @@
 from LinkedList import *
+import jwt
 
 
 class Blockchain:
